make useEffects at same level
