+++ hat/assets/js/apps/Iaso/domains/registry/components/Instances.tsx
@@ -3,7 +3,6 @@ import { Column, useRedirectToReplace } from 'bluesquare-components';
 import React, {
     FunctionComponent,
     useCallback,
-    useEffect,
     useMemo,
     useState,
 } from 'react';
@@ -18,9 +17,8 @@ import { ColumnSelect } from '../../instances/components/ColumnSelect';
 import { OrgunitType } from '../../orgUnits/types/orgunitTypes';
 import { INSTANCE_METAS_FIELDS, defaultSorted } from '../config';
 import { useGetEmptyInstanceOrgUnits } from '../hooks/useGetEmptyInstanceOrgUnits';
-import { useGetForms } from '../hooks/useGetForms';
+import { DropdownOptions } from '../hooks/useGetForms';
 import { useGetInstanceApi, useGetInstances } from '../hooks/useGetInstances';
-import { useGetOrgUnitType } from '../hooks/useGetOrgUnitType';
 import MESSAGES from '../messages';
 import { RegistryParams } from '../types';
 import { OrgunitTypeRegistry } from '../types/orgunitTypes';
@@ -31,31 +29,22 @@ type Props = {
     isLoading: boolean;
     subOrgUnitTypes: OrgunitTypeRegistry[];
     params: RegistryParams;
+    currentType?: OrgunitTypeRegistry;
+    formsList?: DropdownOptions<string>[];
+    isFetchingForms: boolean;
 };
 
 export const Instances: FunctionComponent<Props> = ({
     isLoading,
     subOrgUnitTypes,
     params,
+    currentType,
+    formsList,
+    isFetchingForms,
 }) => {
     const redirectToReplace = useRedirectToReplace();
     const [tableColumns, setTableColumns] = useState<Column[]>([]);
-    const { formIds, tab } = params;
-    const currentType: OrgunitTypeRegistry | undefined = useMemo(() => {
-        if (subOrgUnitTypes.length > 0) {
-            if (tab) {
-                const existingType: OrgunitTypeRegistry | undefined =
-                    subOrgUnitTypes.find(subType => `${subType.id}` === tab);
-                return existingType || subOrgUnitTypes[0];
-            }
-            return subOrgUnitTypes[0];
-        }
-        return undefined;
-    }, [subOrgUnitTypes, tab]);
-    const { data: orgunitTypeDetail } = useGetOrgUnitType(currentType?.id);
-    const { data: formsList, isFetching: isFetchingForms } = useGetForms({
-        orgUnitTypeIds: currentType?.id,
-    });
+    const { formIds } = params;
 
     const { url: apiUrl } = useGetInstanceApi(params, currentType?.id, 'VALID');
     const { data, isFetching: isFetchingList } = useGetInstances(
@@ -93,28 +82,6 @@ export const Instances: FunctionComponent<Props> = ({
     const currentForm: Form | undefined = useMemo(() => {
         return formsList?.find(f => `${f.value}` === formIds)?.original;
     }, [formIds, formsList]);
-
-    useEffect(() => {
-        if (
-            formsList &&
-            (formsList?.length ?? 0) >= 0 &&
-            !isFetchingForms &&
-            !formIds &&
-            orgunitTypeDetail
-        ) {
-            const selectedForm =
-                orgunitTypeDetail.reference_forms.length > 0
-                    ? `${orgunitTypeDetail.reference_forms[0].id}`
-                    : formsList[0].value;
-            const newParams = {
-                ...params,
-                formIds: selectedForm,
-            };
-            redirectToReplace(baseUrls.registry, newParams);
-        }
-        // only prselect a form if forms list contain an element and params is empty
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, [formsList, isFetchingForms, orgunitTypeDetail]);
 
     return (
         <Box>
@@ -171,6 +138,7 @@ export const Instances: FunctionComponent<Props> = ({
                                         instanceMetasFields={
                                             INSTANCE_METAS_FIELDS
                                         }
+                                        // eslint-disable-next-line react/no-unstable-nested-components
                                         getActionCell={settings => (
                                             <ActionCell settings={settings} />
                                         )}
